# services/auth/api/swagger.yaml
basePath: /
definitions:
  http.accountResponse:
    properties:
      created_at:
        example: "2026-10-04T00:00:00Z"
        format: date-time
        type: string
      email:
        example: user@example.com
        format: email
        type: string
      email_verified:
        example: false
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      status:
        example: active
        type: string
    required:
    - created_at
    - email
    - email_verified
    - id
    - status
    type: object
  http.errorDetail:
    properties:
      code:
        example: invalid_request
        type: string
      message:
        example: Invalid email or password
        type: string
      request_id:
        example: c6b8f3a1-4d2e-4f7b-9c1a-2e5f8a9b0c1d
        type: string
    required:
    - code
    - message
    - request_id
    type: object
  http.errorResponse:
    properties:
      error:
        $ref: '#/definitions/http.errorDetail'
    required:
    - error
    type: object
  http.healthResponse:
    properties:
      status:
        example: ok
        type: string
    required:
    - status
    type: object
  http.loginRequest:
    properties:
      email:
        example: user@example.com
        format: email
        maxLength: 254
        type: string
      password:
        example: ExamplePassword123
        maxLength: 128
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
  http.loginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJFZERTQSI...
        type: string
      account:
        $ref: '#/definitions/http.accountResponse'
      expires_in:
        example: 600
        type: integer
      token_type:
        example: Bearer
        type: string
    required:
    - access_token
    - account
    - expires_in
    - token_type
    type: object
  http.meResponseEnvelope:
    properties:
      account:
        $ref: '#/definitions/http.accountResponse'
    type: object
  http.registerRequest:
    properties:
      email:
        example: user@example.com
        format: email
        maxLength: 254
        type: string
      password:
        example: securePassword123!
        maxLength: 128
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  http.registerResponse:
    properties:
      account:
        $ref: '#/definitions/http.accountResponse'
    required:
    - account
    type: object
  token.JWK:
    properties:
      crv:
        type: string
      kid:
        type: string
      kty:
        type: string
      use:
        type: string
      x:
        type: string
    type: object
  token.JWKS:
    properties:
      keys:
        items:
          $ref: '#/definitions/token.JWK'
        type: array
    type: object
info:
  contact: {}
  description: |-
    Authentication service for Gatekeeper.
    Implements AUTH-01 registration and AUTH-02 authentication (login, access JWT, sessions, profile, logout, JWKS).
  title: Gatekeeper Auth Service API
  version: 0.2.0
paths:
  /.well-known/jwks.json:
    get:
      description: Returns the active and archived public signing keys for asymmetric
        Ed25519 token verification without disclosing private key material.
      produces:
      - application/json
      responses:
        "200":
          description: Public JWKS key set
          schema:
            $ref: '#/definitions/token.JWKS'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Get public JSON Web Key Set (JWKS)
      tags:
      - auth
  /healthz:
    get:
      description: Returns 200 OK if the HTTP service process is running and not deadlocked.
      produces:
      - application/json
      responses:
        "200":
          description: Process is healthy
          schema:
            $ref: '#/definitions/http.healthResponse'
      summary: Liveness probe
      tags:
      - probes
  /readyz:
    get:
      description: Returns 200 OK if the service and database are ready to accept
        traffic.
      produces:
      - application/json
      responses:
        "200":
          description: Service and database are ready
          schema:
            $ref: '#/definitions/http.healthResponse'
        "503":
          description: Database connection ping failed or timed out
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Readiness probe
      tags:
      - probes
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates email and password, creates a persistent session,
        and returns a signed Ed25519 access JWT.
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authentication successful
          schema:
            $ref: '#/definitions/http.loginResponse'
        "400":
          description: Malformed JSON or excessive payload
          schema:
            $ref: '#/definitions/http.errorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/http.errorResponse'
        "413":
          description: Payload too large (> 4 KiB)
          schema:
            $ref: '#/definitions/http.errorResponse'
        "415":
          description: Unsupported media type
          schema:
            $ref: '#/definitions/http.errorResponse'
        "429":
          description: Rate limit exceeded (too many login attempts)
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "503":
          description: Database unavailable
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Authenticate user and issue access JWT
      tags:
      - auth
  /v1/auth/logout:
    post:
      description: 'Revokes the active session associated with the bearer access token.
        Idempotent: repeating logout with an unexpired token returns 204 if already
        revoked.'
      responses:
        "204":
          description: Session successfully revoked
        "401":
          description: Unauthorized or session does not exist
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "503":
          description: Database unavailable
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Log out current session
      tags:
      - auth
  /v1/auth/me:
    get:
      description: Retrieves the authenticated account information using verified
        token claims and persisted active session state.
      produces:
      - application/json
      responses:
        "200":
          description: Account profile details
          schema:
            $ref: '#/definitions/http.meResponseEnvelope'
        "401":
          description: Unauthorized or session revoked/expired
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "503":
          description: Database unavailable
          schema:
            $ref: '#/definitions/http.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new registered account with an Argon2id password hash.
        Body is limited to 4 KiB and unknown fields are rejected.
      parameters:
      - description: Registration credentials (email and password)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
          schema:
            $ref: '#/definitions/http.registerResponse'
        "400":
          description: Invalid JSON body, email format, or password length
          schema:
            $ref: '#/definitions/http.errorResponse'
        "409":
          description: Account with this canonical email already exists
          schema:
            $ref: '#/definitions/http.errorResponse'
        "413":
          description: Request body exceeds limit
          schema:
            $ref: '#/definitions/http.errorResponse'
        "415":
          description: Content-Type must be application/json
          schema:
            $ref: '#/definitions/http.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.errorResponse'
        "503":
          description: Service temporarily unavailable, please retry
          schema:
            $ref: '#/definitions/http.errorResponse'
      summary: Register a new account
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"